Move `index_tokens` function to module `token`
